api/requirements.txt
Flask==3.1.0
requests==2.32.3
flask_swagger_ui==4.11.1
python-dotenv==1.0.1
gunicorn
